==> KDT_AI_2/4_FastText.ipynb <==
**Word Embedding 기법을 추가하자**  
**torch==1.9**  
**torchtext==0.10.1**  
**리스트의 각단어들 길이 통일이후 FastText 로 WordEmbedding한다**

Word Embdedding 을할시   
train 입력의 shape 는 (65863, 121, 100)  
test 입력의 shape 는 (13491, 121, 100)  

# !pip install torch==1.9
# !pip install torchtext==0.10.1
import torch
import numpy as np
import pandas as pd
import seaborn as sns
from tqdm import tqdm
from konlpy.tag import Okt  
from collections import Counter
import torch.nn.functional as F
from gensim.models import FastText

train_data = pd.read_excel("./middle/after_tokenized_train.xlsx")[["tokenized_text","label"]]
test_data = pd.read_excel("./middle/after_tokenized_test.xlsx")[["tokenized_text","label"]]

okt=Okt()  
with open('./middle/all_speech_text.txt', 'r', encoding='utf-8') as f:
    text = f.read()

all_morphs = okt.morphs(text)
counted_morphs = Counter(all_morphs)

clean_train_text = [ eval(text) for text in tqdm(train_data['tokenized_text'])]
clean_test_text = [ eval(text) for text in tqdm(test_data['tokenized_text'])]
clean_full_text = clean_train_text + clean_test_text

fast_text_model = FastText(vector_size=100, window=3, min_count=1, sentences=clean_full_text, epochs=20)

fast_text_model.wv.most_similar("이명박")

tokens = list(counted_morphs.keys())
token_to_index = {token: i+1 for i, token in enumerate(tokens)}
index_to_token = {index: token for token, index in token_to_index.items()}
index_to_token[0] = '<unknown>'

def token_to_index_list(token_list):
    token_list = eval(token_list)
    ret = []
    for token in token_list:
        ret.append(token_to_index[token])
    return ret

train_data["indexed_token"] = train_data["tokenized_text"].map(token_to_index_list)
test_data["indexed_token"] = test_data["tokenized_text"].map(token_to_index_list)

indexed_tokens_list_test = [tokens for tokens in test_data['indexed_token']]
indexed_tokens_list_train = [tokens for tokens in train_data['indexed_token']]

train_data.head()

def pad_sequences(sequences, max_len=None):
    if max_len is None:
        max_len = max([len(seq) for seq in sequences])
    padded_seqs = torch.zeros((len(sequences), max_len)).long()
    for i, seq in enumerate(sequences):
        end = min(len(seq), max_len)
        padded_seqs[i, :end] = torch.LongTensor(seq[:end])
    return padded_seqs

train_input_list = pad_sequences(indexed_tokens_list_train,121)
test_input_list = pad_sequences(indexed_tokens_list_test,121)
train_answer_list = [int(label) for label in train_data['label']]
train_answer_list = torch.tensor(train_answer_list)

print(train_input_list.shape)
print(test_input_list.shape)
print(len(train_answer_list))

# tokens = list(counted_morphs.keys())
# token_to_index = {token: i+1 for i, token in enumerate(tokens)}
# index_to_token = {index: token for token, index in token_to_index.items()}
# index_to_token[0] = '<unkown>'

final_train_input = []

for tmp_input_list in train_input_list:
    ret = []
    for tmp in tmp_input_list:
        tmp_to_word = index_to_token[int(tmp)]
        if tmp_to_word == "<unknown>":
            arr = np.zeros(100)
        arr = fast_text_model.wv[tmp_to_word]
        arr = arr.tolist()
        ret.append(arr)
    final_train_input.append(ret)
    
final_train_input = np.array(final_train_input)

print(final_train_input.shape)

final_test_input = []

for tmp_input_list in test_input_list:
    ret = []
    for tmp in tmp_input_list:
        tmp_to_word = index_to_token[int(tmp)]
        if tmp_to_word == "<unknown>":
            arr = np.zeros(100)
        arr = fast_text_model.wv[tmp_to_word]
        arr = arr.tolist()
        ret.append(arr)
    final_test_input.append(ret)
    
final_test_input = np.array(final_test_input)

print(final_test_input.shape)

final_test_input = torch.Tensor(final_test_input)
final_train_input = torch.Tensor(final_train_input)

torch.save(final_train_input, './processed_data/fasttext_train_input.pt')
torch.save(final_test_input,  './processed_data/fast_text_test_input.pt')
torch.save(train_answer_list, './processed_data/train_label.pt')
