==> KDT_AI_2/6_submit.ipynb <==
https://colab.research.google.com/drive/13eTxQzJaPzpkF2T6DWMdYzJLPQENZRZt

**로컬 환경이 안되니 colab GPU 환경에서 훈련시킴**

import numpy as np
import pandas as pd

submit = np.load("./output/final_output.npy")

submission_df = pd.read_csv("./output/jhkim_1.csv")
submission_df.head()

submission_df['label'] = submit_list

submission_df.to_csv("./output/jhkim_final.csv",index=False)
